requirements.txt
altgraph==0.17.4
macholib==1.16.3
packaging==24.0
pyinstaller==6.5.0
pyinstaller-hooks-contrib==2024.3
pypandoc_binary==1.13
PySide6==6.6.3
PySide6_Addons==6.6.3
PySide6_Essentials==6.6.3
shiboken6==6.6.3
